+++ src/fires.hpp
@@ -7,15 +7,15 @@
 #include "matrix.hpp"
 
 struct Fire {
-  unsigned int width;
-  unsigned int height;
+  size_t width;
+  size_t height;
 
   Matrix<bool> burned_layer;
 
-  std::vector<std::pair<unsigned int, unsigned int>> burned_ids;
+  std::vector<std::pair<size_t, size_t>> burned_ids;
 
   // Positions in burned_ids where a new step starts, empty if the fire was not simulated
-  std::vector<unsigned int> burned_ids_steps;
+  std::vector<size_t> burned_ids_steps;
 
   bool operator==(const Fire& other) const {
     return width == other.width && height == other.height &&
@@ -23,15 +23,15 @@ struct Fire {
   }
 };
 
-Fire read_fire(unsigned int width, unsigned int height, std::string filename);
+Fire read_fire(size_t width, size_t height, std::string filename);
 
 /* Type and function useful for comparing fires */
 
 struct FireStats {
-  unsigned int counts_veg_matorral;
-  unsigned int counts_veg_subalpine;
-  unsigned int counts_veg_wet;
-  unsigned int counts_veg_dry;
+  size_t counts_veg_matorral;
+  size_t counts_veg_subalpine;
+  size_t counts_veg_wet;
+  size_t counts_veg_dry;
 };
 
 FireStats get_fire_stats(const Fire& fire, const Landscape& landscape);
